requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
black==20.8b1
certifi==2020.11.8
cffi==1.14.3
cisco-gnmi==1.0.11
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
cryptography==3.2.1
future==0.18.2
grpcio==1.33.2
grpcio-tools==1.33.2
h11==0.11.0
httpcore==0.12.1
httpx==0.16.1
idna==2.10
lxml==4.6.1
mypy-extensions==0.4.3
ncclient==0.6.9
netmiko==3.3.2
ntc-templates==1.6.0
paramiko==2.7.2
pathspec==0.8.1
protobuf==3.14.0
pyang==2.4.0
pycparser==2.20
Pygments==2.7.2
PyNaCl==1.4.0
pyserial==3.4
regex==2020.11.13
rfc3986==1.4.0
rich==9.2.0
scp==0.13.3
six==1.15.0
sniffio==1.2.0
tenacity==6.2.0
textfsm==1.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
xmltodict==0.12.0
